+++ src/index.ts
@@ -135,14 +135,14 @@ export class AeroflyMission {
     tags: string[];
 
     /**
-     * @property {boolean} isFeatured makes this mission pop up in "Challenges"
+     * @property {?boolean} isFeatured makes this mission pop up in "Challenges"
      */
-    isFeatured: boolean;
+    isFeatured: boolean | null;
 
     /**
-     * @property {number|undefined} difficulty values between 0.00 and 2.00 have been encountered, but they seem to be without limit
+     * @property {?number} difficulty values between 0.00 and 2.00 have been encountered, but they seem to be without limit
      */
-    difficulty: number | undefined;
+    difficulty: number | null;
 
     /**
      * @property {"taxi"|"takeoff"|"cruise"|"approach"|"landing"|"winch_launch"|"aerotow"} flightSetting of aircraft, like "taxi", "cruise"
@@ -170,14 +170,14 @@ export class AeroflyMission {
     destination: AeroflyMissionPosition;
 
     /**
-     * @property {number} in meters
+     * @property {?number} distance in meters
      */
-    distance: number;
+    distance: number | null;
 
     /**
-     * @property {number} in seconds
+     * @property {?number} duration in seconds
      */
-    duration: number;
+    duration: number | null;
 
     /**
      * @property {?AeroflyMissionTargetPlane} finish as finish condition
@@ -200,15 +200,15 @@ export class AeroflyMission {
      * @param {string} [additionalAttributes.description] text, mission briefing, etc
      * @param {AeroflyLocalizedText[]} [additionalAttributes.localizedTexts] translations for title and description
      * @param {string[]} [additionalAttributes.tags]
-     * @param {boolean} [additionalAttributes.isFeatured] makes this mission pop up in "Challenges"
-     * @param {number|undefined} [additionalAttributes.difficulty] values between 0.00 and 2.00 have been encountered, but they seem to be without limit
+     * @param {?boolean} [additionalAttributes.isFeatured] makes this mission pop up in "Challenges"
+     * @param {?number} [additionalAttributes.difficulty] values between 0.00 and 2.00 have been encountered, but they seem to be without limit
      * @param {"taxi"|"takeoff"|"cruise"|"approach"|"landing"|"winch_launch"|"aerotow"} [additionalAttributes.flightSetting] of aircraft, like "taxi", "cruise"
      * @param {{name:string,livery:string,icao:string}} [additionalAttributes.aircraft] for this mission
      * @param {string} [additionalAttributes.callsign] of aircraft, uppercased
      * @param {object} [additionalAttributes.origin] position of aircraft, as well as name of starting airport. Position does not have match airport.
      * @param {object} [additionalAttributes.destination] position of aircraft, as well as name of destination airport. Position does not have match airport.
-     * @param {number} [additionalAttributes.distance] in meters
-     * @param {number} [additionalAttributes.duration] in seconds
+     * @param {?number} [additionalAttributes.distance] in meters
+     * @param {?number} [additionalAttributes.duration] in seconds
      * @param {?AeroflyMissionTargetPlane} [additionalAttributes.finish] as finish condition
      * @param {AeroflyMissionConditions} [additionalAttributes.conditions] like time and weather for mission
      * @param {AeroflyMissionCheckpoint[]} [additionalAttributes.checkpoints] form the actual flight plan
@@ -219,8 +219,8 @@ export class AeroflyMission {
             description = "",
             localizedTexts = [],
             tags = [],
-            isFeatured = false,
-            difficulty = undefined,
+            isFeatured = null,
+            difficulty = null,
             flightSetting = "taxi",
             aircraft = {
                 name: "c172",
@@ -242,8 +242,8 @@ export class AeroflyMission {
                 dir: 0,
                 alt: 0,
             },
-            distance = 0,
-            duration = 0,
+            distance = null,
+            duration = null,
             finish = null,
             conditions = new AeroflyMissionConditions(),
             checkpoints = [],
@@ -251,15 +251,15 @@ export class AeroflyMission {
             description?: string;
             localizedTexts?: AeroflyLocalizedText[];
             tags?: string[];
-            isFeatured?: boolean;
-            difficulty?: number | undefined;
+            isFeatured?: boolean | null;
+            difficulty?: number | null;
             flightSetting?: AeroflyMissionSetting;
             aircraft?: AeroflyMissionAircraft;
             callsign?: string;
             origin?: AeroflyMissionPosition;
             destination?: AeroflyMissionPosition;
-            distance?: number;
-            duration?: number;
+            distance?: number | null;
+            duration?: number | null;
             finish?: AeroflyMissionTargetPlane | null;
             conditions?: AeroflyMissionConditions;
             checkpoints?: AeroflyMissionCheckpoint[];
@@ -345,27 +345,27 @@ ${this.getLocalizedTextsString()}
         if (this.tags.length > 0) {
             optionalProperties.push(`                <[string8u][tags][ ${this.tags.join(" ")} ]>`);
         }
-        if (this.difficulty !== undefined) {
+        if (this.difficulty !== null) {
             optionalProperties.push(`                <[float64]   [difficulty]         [${this.difficulty}]>`);
         }
-        if (this.isFeatured) {
+        if (this.isFeatured !== null) {
             optionalProperties.push(
                 `                <[bool]      [is_featured]        [${this.isFeatured ? "true" : "false"}]>`,
             );
         }
 
         const moreOptionalProperties = [];
-        if (this.distance) {
+        if (this.distance !== null) {
             moreOptionalProperties.push(
                 `                <[float64]   [distance]           [${this.distance}]> // ${Math.ceil(this.distance / 1000)} km`,
             );
         }
-        if (this.duration) {
+        if (this.duration !== null) {
             moreOptionalProperties.push(
                 `                <[float64]   [duration]           [${this.duration}]> // ${Math.ceil(this.duration / 60)} min`,
             );
         }
-        if (this.finish) {
+        if (this.finish !== null) {
             moreOptionalProperties.push(this.finish.toString());
         }
 
@@ -441,8 +441,8 @@ export class AeroflyMissionConditions {
      * @param {number} [additionalAttributes.turbulenceStrength] 0..1, percentage
      * @param {number} [additionalAttributes.thermalStrength] 0..1, percentage
      * @param {number} [additionalAttributes.visibility] in meters
-     * @param {number?} [additionalAttributes.visibility_sm] in statute miles, will overwrite visibility
-     * @param {number?} [additionalAttributes.temperature] in °C, will overwrite thermalStrength
+     * @param {?number} [additionalAttributes.visibility_sm] in statute miles, will overwrite visibility
+     * @param {?number} [additionalAttributes.temperature] in °C, will overwrite thermalStrength
      * @param {AeroflyMissionConditionsCloud[]} [additionalAttributes.clouds] for the whole flight
      */
     constructor({
@@ -721,9 +721,9 @@ export class AeroflyMissionCheckpoint {
     frequency: number | null;
 
     /**
-     * @property {boolean|undefined} flyOver if waypoint is meant to be flown over
+     * @property {?boolean} flyOver if waypoint is meant to be flown over
      */
-    flyOver: boolean | undefined;
+    flyOver: boolean | null;
 
     /**
      * @param {string} name ICAO code for airport, runway designator, navaid
@@ -738,14 +738,14 @@ export class AeroflyMissionCheckpoint {
      * @param {object} additionalAttributes allows to set additional attributes on creation
      * @param {number} [additionalAttributes.altitude] The height in meters above or below the WGS
      *    84 reference ellipsoid
-     * @param {number?} [additionalAttributes.altitude_feet] The height in feet above or below the WGS
+     * @param {?number} [additionalAttributes.altitude_feet] The height in feet above or below the WGS
      *    84 reference ellipsoid. Will overwrite altitude
      * @param {number} [additionalAttributes.direction] of runway, in degree
-     * @param {number?} [additionalAttributes.slope] of runway
-     * @param {number?} [additionalAttributes.length] of runway, in meters
-     * @param {number?} [additionalAttributes.length_feet] of runway, in feet. Will overwrite length
-     * @param {number?} [additionalAttributes.frequency] of runways or navigational aids, in Hz; multiply by 1000 for kHz, 1_000_000 for MHz
-     * @param {boolean|undefined} [additionalAttributes.flyOver] if waypoint is meant to be flown over
+     * @param {?number} [additionalAttributes.slope] of runway
+     * @param {?number} [additionalAttributes.length] of runway, in meters
+     * @param {?number} [additionalAttributes.length_feet] of runway, in feet. Will overwrite length
+     * @param {?number} [additionalAttributes.frequency] of runways or navigational aids, in Hz; multiply by 1000 for kHz, 1_000_000 for MHz
+     * @param {?boolean} [additionalAttributes.flyOver] if waypoint is meant to be flown over
      */
     constructor(
         name: string,
@@ -760,7 +760,7 @@ export class AeroflyMissionCheckpoint {
             length = null,
             length_feet = null,
             frequency = null,
-            flyOver = undefined,
+            flyOver = null,
         }: {
             altitude?: number;
             altitude_feet?: number | null;
@@ -769,7 +769,7 @@ export class AeroflyMissionCheckpoint {
             length?: number | null;
             length_feet?: number | null;
             frequency?: number | null;
-            flyOver?: boolean | undefined;
+            flyOver?: boolean | null;
         } = {},
     ) {
         this.type = type;
@@ -852,7 +852,7 @@ export class AeroflyMissionCheckpoint {
                 `                        <[float64][frequency][${this.frequency ?? 0}]> // ${this.frequency_string}`,
             );
         }
-        if (this.flyOver !== undefined) {
+        if (this.flyOver !== null) {
             optionalProperties.push(`                        <[bool][fly_over][${this.flyOver ? "true" : "false"}]>`);
         }